Error nhieu qua

+++ src/main/java/com/nhom27/nhatkykhambenh/controller/DonThuocController.java
@@ -1,8 +1,16 @@
 package com.nhom27.nhatkykhambenh.controller;
 
 import com.nhom27.nhatkykhambenh.dto.DonThuocDTO;
+import com.nhom27.nhatkykhambenh.dto.XetNghiemDTO;
 import com.nhom27.nhatkykhambenh.exception.SaveDataException;
+import com.nhom27.nhatkykhambenh.mapper.DonThuocMapper;
+import com.nhom27.nhatkykhambenh.mapper.XetNghiemMapper;
+import com.nhom27.nhatkykhambenh.model.DonThuoc;
+import com.nhom27.nhatkykhambenh.model.XetNghiem;
 import com.nhom27.nhatkykhambenh.service.implementation.DonThuocService;
+import com.nhom27.nhatkykhambenh.service.interfaces.IChiTietKhamBenhService;
+import com.nhom27.nhatkykhambenh.service.interfaces.IDonThuocService;
+import com.nhom27.nhatkykhambenh.service.interfaces.IXetNghiemService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -22,19 +30,29 @@ import java.util.List;
 @Controller
 @RequiredArgsConstructor
 public class DonThuocController {
+
     @Autowired
-    private DonThuocService donThuocService;
+    private IDonThuocService donThuocService;
 
-    @GetMapping("/admin/donthuoc")
+    @Autowired
+    private DonThuocMapper donThuocMapper;
+
+    @Autowired
+    private IChiTietKhamBenhService chiTietKhamBenhService;
+
+    @GetMapping("/admin/khambenh/chitiet/donthuoc")
     public String GetListDonThuoc(Model model,
-                                  @RequestParam(defaultValue = "0") int page,
-                                  @RequestParam(defaultValue = "5") int size,
-                                  @RequestParam(defaultValue = "") String query) {
+                                   @RequestParam(defaultValue = "0") int page,
+                                   @RequestParam(defaultValue = "5") int size,
+                                   @RequestParam(defaultValue = "") String query,
+                                   @RequestParam Integer maChiTietKhamBenh) {
         Pageable pageable = PageRequest.of(page, size);
-        Page<DonThuocDTO> donThuocPage = donThuocService.getDSDonThuoc(pageable,query);
-        int count = 0;
+        Page<DonThuoc> donThuocPage = donThuocService.getDSDonThuoc(pageable, query, maChiTietKhamBenh);
+        List<DonThuocDTO> donThuocDTOList = donThuocMapper.toDonThuocDtoList(donThuocPage.getContent());
 
-        model.addAttribute("dsDonThuoc", donThuocPage.getContent());
+
+        model.addAttribute("dsXetNghiem", donThuocDTOList);
+        model.addAttribute("maChiTietKhamBenh", maChiTietKhamBenh);
         model.addAttribute("currentPage", page);
         model.addAttribute("pageSize", size);
         model.addAttribute("totalPages", donThuocPage.getTotalPages());
@@ -49,64 +67,86 @@ public class DonThuocController {
         model.addAttribute("endItem", endItem);
         model.addAttribute("currentCount", endItem - startItem + 1);
 
-        System.out.println("currentPage: " + page);
-        System.out.println("currentPage: " + size);
-        System.out.println("currentPage: " + donThuocPage.getTotalPages());
-        System.out.println("currentPage: " + donThuocPage.getTotalElements());
-        System.out.println("currentPage: " + query);
-        return "admin/donthuoc/listDonThuoc";
+        return "admin/khambenh/listDonThuoc";
     }
 
-    @GetMapping("/admin/donthuoc/add")
-    public String addDonThuocForm(Model model) {
+    @GetMapping("/admin/khambenh/chitiet/donthuoc/add")
+    public String addDonThuocForm(Model model, @RequestParam Integer maChiTietKhamBenh) {
         DonThuocDTO donThuocDTO = new DonThuocDTO();
+        donThuocDTO.setMaChiTietKhamBenh(maChiTietKhamBenh);
         model.addAttribute("donthuoc", donThuocDTO);
+        model.addAttribute("maChiTietKhamBenh2", maChiTietKhamBenh);
+
         return "admin/donthuoc/addDonThuoc";
     }
 
 
-    @GetMapping("/admin/donthuoc/update")
-    public String updateDonThuocForm(@RequestParam("id") Integer id, Model model) {
-        DonThuocDTO donThuocDTO = donThuocService.findById(id);
+    @GetMapping("/admin/khambenh/chitiet/donthuoc/update")
+    public String updateDonThuocForm(@RequestParam Integer maDonThuoc,
+                                     @RequestParam Integer maChiTietKhamBenh,
+                                     Model model) {
+        DonThuoc donThuoc = donThuocService.findById(maDonThuoc);
+        DonThuocDTO donThuocDTO = donThuocMapper.toDonThuocDTO(donThuoc);
+        donThuocDTO.setMaChiTietKhamBenh(maChiTietKhamBenh);
+
         model.addAttribute("donthuoc", donThuocDTO);
-        return "admin/donthuoc/addDonThuoc";
+        model.addAttribute("maChiTietKhamBenh2", maChiTietKhamBenh);
+        return "admin/khambenh/addDonThuoc";
     }
 
-    @PostMapping("/admin/donthuoc/save")
-    public String saveDonThuoc(@ModelAttribute("donthuoc") DonThuocDTO donthuoc,
+    @PostMapping("/admin/khambenh/chitiet/donthuoc/save")
+    public String saveDonThuoc(@ModelAttribute("donthuoc") DonThuocDTO donThuocDTO,
                                BindingResult bindingResult,
-                               Model model) {
+                               Model model,
+                               @RequestParam(required = false) Integer maChiTietKhamBenh,
+                               @RequestParam(required = false) Integer maDonThuoc,
+                               RedirectAttributes redirectAttributes) {
         if (bindingResult.hasErrors()) {
-            return "admin/donthuoc/addDonThuoc";
+            return "admin/khambenh/addDonThuoc";
         }
         try {
-            donThuocService.saveDonThuoc(donthuoc);
-            return "redirect:/admin/donthuoc";
+            if (maDonThuoc != null) {
+                donThuocDTO.setMaDonThuoc(maDonThuoc);
+            }
+            DonThuoc donThuoc = donThuocMapper.toDonThuoc(donThuocDTO);
+            donThuocService.saveDonThuoc(donThuoc,maChiTietKhamBenh);
+            redirectAttributes.addAttribute("maChiTietKhamBenh", maChiTietKhamBenh);
+            return "redirect:/admin/khambenh/chitiet/donthuoc";
         } catch (SaveDataException e) {
             model.addAttribute("error", e.getMessage());
-            return "admin/donthuoc/addDonThuoc";
+            return "admin/khambenh/addDonThuoc";
         }
     }
 
-    @PostMapping("/admin/donthuoc/delete")
-    public String deleteDonThuoc(@RequestParam("maDonThuoc") Integer maDonThuoc, RedirectAttributes redirectAttributes) {
+    @PostMapping("/admin/khambenh/chitiet/donthuoc/delete")
+    public String deleteDonThuoc(@RequestParam("maDonThuoc") Integer maDonThuoc,@RequestParam("maChiTietKhamBenh") Integer maChiTietKhamBenh,
+                                 RedirectAttributes redirectAttributes) {
         try {
             donThuocService.deleteById(maDonThuoc);
             redirectAttributes.addFlashAttribute("success", "Xóa thành công");
         } catch (Exception e) {
             redirectAttributes.addFlashAttribute("error", "Lỗi!! Xóa thông tin Đơn Thuốc thất bại");
         }
-        return "redirect:/admin/donthuoc";
+        redirectAttributes.addAttribute("maChiTietKhamBenh", maChiTietKhamBenh);
+        return "redirect:/admin/khambenh/chitiet/donthuoc";
     }
 
-    @PostMapping("/admin/donthuoc/deleteall")
-    public String deleteAllByIds(@RequestParam("selectedIds") List<Integer> ids, RedirectAttributes redirectAttributes) {
+    @PostMapping("/admin/khambenh/chitiet/donthuoc/deleteall")
+    public String deleteAllByIds(@RequestParam(value = "selectedIds", required = false) List<Integer> ids,
+                                 @RequestParam("maChiTietKhamBenh") Integer maChiTietKhamBenh,
+                                RedirectAttributes redirectAttributes) {
+        if (ids == null || ids.isEmpty()) {
+            redirectAttributes.addFlashAttribute("error", "Không có mục nào được chọn để xóa.");
+            redirectAttributes.addAttribute("maChiTietKhamBenh", maChiTietKhamBenh);  // Thêm maKhamBenh vào redirect
+            return "redirect:/admin/khambenh/chitiet/donthuoc";
+        }
         try {
             donThuocService.deleteAllByIds(ids);
             redirectAttributes.addFlashAttribute("success", "Xóa thành công các mục đã chọn.");
         } catch (Exception e) {
-            redirectAttributes.addFlashAttribute("error", "Có lỗi xảy ra khi xóa.");
+            redirectAttributes.addFlashAttribute("error", "Có lỗi xảy ra khi xóa các mục đã chọn.");
         }
-        return "redirect:/admin/donthuoc";
+        redirectAttributes.addAttribute("maChiTietKhamBenh", maChiTietKhamBenh);
+        return "redirect:/admin/khambenh/chitiet/donthuoc";
     }
 }
